% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{error}
\alias{error}
\alias{MAE}
\alias{Error}
\alias{RMSE}
\alias{frobenius}
\alias{MSE}
\title{Error Calculation}
\usage{
MSE(true, predicted, na.rm = TRUE)
RMSE(true, predicted, na.rm = TRUE)
MAE(true, predicted, na.rm = TRUE)
frobenius(true, predicted, na.rm = TRUE)
}
\arguments{
\item{true}{true values.}

\item{predicted}{predicted values}

\item{na.rm}{ignore missing values.}
}
\value{
The error value.
}
\description{
Calculate the mean absolute error (MAE), mean square error (MSE),
root mean square error (RMSE) and for matrices also the Frobenius norm (identical to RMSE).
}
\details{
Frobenius norm requires matrices.
}
\examples{
true <- rnorm(10)
predicted <- rnorm(10)

MAE(true, predicted)
MSE(true, predicted)
RMSE(true, predicted)

true <- matrix(rnorm(9), nrow = 3)
predicted <- matrix(rnorm(9), nrow = 3)

frobenius(true, predicted)
}
\seealso{
Other evaluation:
\code{\link[=calcPredictionAccuracy]{calcPredictionAccuracy()}},
\code{\link[=evaluate]{evaluate()}},
\code{\link{evaluationResultList-class}},
\code{\link{evaluationResults-class}},
\code{\link[=evaluationScheme]{evaluationScheme()}},
\code{\link{evaluationScheme-class}},
\code{\link[=plot]{plot()}}
}
\concept{evaluation}
