% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{evaluationResultList-class}
\alias{evaluationResultList-class}
\alias{coerce,list,evaluationResultList-method}
\alias{avg,evaluationResultList-method}
\alias{[,evaluationResultList,ANY,missing,missing-method}
\alias{show,evaluationResultList-method}
\title{Class "evaluationResultList": Results from Evaluating Multiple Recommender Methods}
\description{
Contains evaluation results for several runs of multiple recommender methods, represented as confusion matrices. The models used for each run may also be available.
}
\seealso{
\code{\link{evaluate}},
	\code{\linkS4class{evaluationResults}}.

Other evaluation:
\code{\link[=calcPredictionAccuracy]{calcPredictionAccuracy()}},
\code{\link{error}},
\code{\link[=evaluate]{evaluate()}},
\code{\link{evaluationResults-class}},
\code{\link[=evaluationScheme]{evaluationScheme()}},
\code{\link{evaluationScheme-class}},
\code{\link[=plot]{plot()}}
}
\concept{evaluation}
\keyword{classes}
\section{Objects from the Class}{
Objects are created by \code{evaluate}.
}
\section{Slots}{
	 \describe{
    \item{\code{.Data}:}{Object of class \code{"list"}: a list of
	\code{"evaluationResults"}.}
  }
}
\section{Extends}{
Class \code{"\linkS4class{list}"}, from data part.
%Class \code{"\linkS4class{vector}"}, by class "list", distance 2.
%Class \code{"\linkS4class{listOrNull}"}, by class "list", distance 2.
}
\section{Methods}{
  \describe{
    \item{avg}{\code{signature(x = "evaluationResultList")}: returns a
	list of average confusion matrices.}
    \item{[}{\code{signature(x = "evaluationResultList", i = "ANY", j = "missing", drop = "missing")}}
%    \item{plot}{\code{signature(x = "evaluationResultList")}: ... }
    \item{coerce}{\code{signature(from = "list", to = "evaluationResultList")}}
	\item{show}{\code{signature(object = "evaluationResultList")}}
	 }
}
