% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{evaluationResults-class}
\alias{evaluationResults-class}
\alias{confusionMatrix-class}
\alias{avg}
\alias{avg,evaluationResults-method}
\alias{getConfusionMatrix}
\alias{getConfusionMatrix,evaluationResults-method}
\alias{getResults}
\alias{getResults,evaluationResults-method}
\alias{getModel}
\alias{getModel,evaluationResults-method}
\alias{getRuns}
\alias{getRuns,evaluationResults-method}
\alias{show,evaluationResults-method}
\title{Class "evaluationResults": Results of the Evaluation of a Single Recommender Method}
\description{
Contains evaluation results for several runs of the same recommender method, represented as confusion matrices. The model used for each run may also be available.
}
\seealso{
\code{\link{evaluate}}

Other evaluation:
\code{\link[=calcPredictionAccuracy]{calcPredictionAccuracy()}},
\code{\link{error}},
\code{\link[=evaluate]{evaluate()}},
\code{\link{evaluationResultList-class}},
\code{\link[=evaluationScheme]{evaluationScheme()}},
\code{\link{evaluationScheme-class}},
\code{\link[=plot]{plot()}}
}
\concept{evaluation}
\keyword{classes}
\section{Objects from the Class}{
Objects are created by \code{evaluate}.
}
\section{Slots}{
	 \describe{
    \item{\code{results}:}{Object of class \code{"list"}: contains
	objects of class \code{"ConfusionMatrix"}, one for each run specified
	in the used evaluation scheme.}
  }
}
\section{Methods}{
  \describe{
    \item{avg}{\code{signature(x = "evaluationResults")}: returns evaluation metrics averaged across cross-validation folds. }
    \item{getConfusionMatrix}{\code{signature(x = "evaluationResults")}:
	Deprecated. Use \code{getResults()}.}
    \item{getResults}{\code{signature(x = "evaluationResults")}:
	returns a list of evaluation metrics with one element for each cross-validation fold.}
    \item{getModel}{\code{signature(x = "evaluationResults")}: returns a
	list of recommender models used, if available. }
    \item{getRuns}{\code{signature(x = "evaluationResults")}: returns
	the number of runs/number of confusion matrices.}
%    \item{plot}{\code{signature(x = "evaluationResults")}: plot }
    \item{show}{\code{signature(object = "evaluationResults")} }
	 }
}
