% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{getList}
\alias{getList}
\alias{getList,binaryRatingMatrix-method}
\alias{getList,realRatingMatrix-method}
\alias{getList,topNList-method}
\alias{getData.frame}
\alias{getData.frame,ratingMatrix-method}
\title{List and Data.frame Representation for Recommender Matrix Objects}
\usage{
getList(from, ...)
\S4method{getList}{realRatingMatrix}(from, decode = TRUE, ratings = TRUE, ...)
\S4method{getList}{binaryRatingMatrix}(from, decode = TRUE, ...)
\S4method{getList}{topNList}(from, decode = TRUE, ...)

getData.frame(from, ...)
\S4method{getData.frame}{ratingMatrix}(from, decode = TRUE, ratings = TRUE, ...)
}
\arguments{
\item{from}{object to be represented as a list.}

\item{...}{further arguments (currently unused).}

\item{decode}{use item names or item IDs (column numbers) for items?}

\item{ratings}{include ratings in the list or data.frame?}
}
\value{
Returns a list or a data.frame.
}
\description{
Create a list or data.frame representation for various objects
used in \pkg{recommenderlab}. These functions are used in addition to
available coercion to allow for parameters like \code{decode}.
}
\details{
Lists have one vector with items (and ratings) per user. The
data.frame has one row per rating with the user in the first column,
the item as the second and the rating as the third.
}
\examples{
data(Jester5k)

getList(Jester5k[1,])
getData.frame(Jester5k[1,])
}
\seealso{
\code{\linkS4class{binaryRatingMatrix}},
\code{\linkS4class{realRatingMatrix}},
\code{\linkS4class{topNList}}.

Other data preparation:
\code{\link[=normalize]{normalize()}},
\code{\link{sparseNAMatrix}}
}
\concept{data preparation}
