\name{.get_parameters}
\Rdversion{1.1}
\alias{.get_parameters}
\title{
  Helper Functions
}
\description{
Internal helper functions.
}
\usage{
.get_parameters(p, parameter)
}
\arguments{
  \item{p}{default parameters.}
  \item{parameter}{user specified parameters.}
}
%\details{}
\value{
Returns a user specified parameter list completed with default parameters.
}
%\seealso{}
%\examples{}
%\keyword{ ~kwd1 }
%\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
