% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HybridRecommender.R
\name{HybridRecommender}
\alias{HybridRecommender}
\title{Create a Hybrid Recommender}
\usage{
HybridRecommender(..., weights = NULL, aggregation_type = "sum")
}
\arguments{
\item{\dots}{objects of class 'Recommender'.}

\item{weights}{weights for the recommenders. The recommenders are equally
weighted by default.}

\item{aggregation_type}{How are the recommendations aggregated. Options are "sum", "min", and
"max".}
}
\value{
An object of class 'Recommender'.
}
\description{
Creates and combines recommendations using several recommender algorithms.
}
\details{
The hybrid recommender is initialized with a set of pretrained Recommender objects.
    Typically, the algorithms are trained using the same training set. If different
    training sets are used, then, at least the training sets
    need to have the same items in the same order.

    Alternatively, hybrid recommenders can be created using the regular \code{Recommender()}
    interface. Here \code{method} is set to \code{HYBRID} and \code{parameter} contains
    a list with recommenders and weights. The recommenders are a list of recommender algorithms,
    each represented by a list with the elements \code{name} (the recommender method)
    and \code{parameters} (the algorithm parameters). This interface can be used with \code{evaluate()}.


    For creating recommendations (\code{predict}), each recommender algorithm
    is used to create ratings. The individual ratings are combined using
    a weighted sum where missing ratings are ignored. Weights can be specified in \code{weights}.
}
\examples{
data("MovieLense")
MovieLense100 <- MovieLense[rowCounts(MovieLense) >100,]
train <- MovieLense100[1:100]
test <- MovieLense100[101:103]

## mix popular movies with a random recommendations for diversity and
## rerecommend some movies the user liked.
recom <- HybridRecommender(
  Recommender(train, method = "POPULAR"),
  Recommender(train, method = "RANDOM"),
  Recommender(train, method = "RERECOMMEND"),
  weights = c(.6, .1, .3)
  )

recom

getModel(recom)

as(predict(recom, test), "list")

## create a hybrid recommender using the regular Recommender interface.
## This is needed to use hybrid recommenders with evaluate().
recommenders <- list(
  RANDOM = list(name = "POPULAR", param = NULL),
  POPULAR = list(name = "RANDOM", param = NULL),
  RERECOMMEND = list(name = "RERECOMMEND", param = NULL)
)

weights <- c(.6, .1, .3)

recom <- Recommender(train, method = "HYBRID",
  parameter = list(recommenders = recommenders, weights = weights))
recom

as(predict(recom, test), "list")
}
\seealso{
\code{\linkS4class{Recommender}}

Other recommender models:
\code{\link[=Recommender]{Recommender()}},
\code{\link{Recommender-class}}
}
\concept{recommender models}
