% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{internal}
\alias{internal}
\alias{returnRatings}
\alias{internalFunctions}
\alias{getParameters}
\title{Internal Utility Functions}
\usage{
returnRatings(ratings, newdata,
  type = c("topNList", "ratings", "ratingMatrix"),
  n, randomize = NULL, minRating = NA)

getParameters(defaults, parameter)
}
\arguments{
\item{ratings}{a realRatingMatrix.}

\item{newdata}{a realRatingMatrix.}

\item{type}{type of recommendation to return.}

\item{n}{max. number of entries in the top-N list.}

\item{randomize}{randomization factor for producing the top-N list.}

\item{minRating}{do not include ratings less than this.}

\item{defaults}{list with parameters and default values.}

\item{parameter}{list with actual parameters.}
}
\description{
Utility functions used internally by recommender algorithms. See files starting
with \code{RECOM} in the package's \code{R} directory for examples of usage.
}
\details{
\code{returnRatings} is used in the predict function of recommender algorithms
to return different types of recommendations.

\code{getParameters} is a helper function which checks parameters for
consistency and provides default values. Used in the Recommender constructor.
}
\seealso{
Other internal:
\code{\link[=funkSVD]{funkSVD()}}
}
\concept{internal}
