% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MovieLense}
\alias{MovieLense}
\alias{MovieLenseMeta}
\alias{MovieLenseUser}
\title{MovieLense Dataset (100k)}
\format{
The format of \code{MovieLense} is an object of class \code{"realRatingMatrix"}

  The format of \code{MovieLenseMeta} is a data.frame with movie title, year, IMDb URL and indicator variables for 19 genres.

  The format of \code{MovieLenseUser} is a data.frame with user age, sex, occupation and zip code.
}
\source{
GroupLens Research, \url{https://grouplens.org/datasets/movielens/}
}
\usage{
data(MovieLense)
}
\description{
The 100k MovieLense
ratings data set. The data was collected through the MovieLens web site
(movielens.umn.edu) during the seven-month period from September 19th,
1997 through April 22nd, 1998.
The data set contains about 100,000 ratings (1-5)
from 943 users on 1664 movies. Movie and user metadata is also provided in \code{MovieLenseMeta} and \code{MovieLenseUser}.
}
\examples{
data(MovieLense)
MovieLense

## look at the first few ratings of the first user
head(as(MovieLense[1,], "list")[[1]])

## visualize part of the matrix
image(MovieLense[1:100,1:100])

## number of ratings per user
hist(rowCounts(MovieLense))

## number of ratings per movie
hist(colCounts(MovieLense))

## mean rating (averaged over users)
mean(rowMeans(MovieLense))

## available movie meta information
head(MovieLenseMeta)

## available user meta information
head(MovieLenseUser)
}
\references{
Herlocker, J., Konstan, J., Borchers, A., Riedl, J.. An Algorithmic
Framework for Performing Collaborative Filtering. Proceedings of the
1999 Conference on Research and Development in Information
Retrieval. Aug. 1999.
}
\seealso{
Other datasets:
\code{\link{Jester5k}},
\code{\link{MSWeb}}
}
\concept{datasets}
\keyword{datasets}
