% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MSWeb}
\alias{MSWeb}
\title{Anonymous web data from www.microsoft.com}
\format{
The format is: Formal class \code{"binaryRatingMatrix"}.
}
\source{
Asuncion, A., Newman, D.J. (2007). UCI Machine Learning Repository, Irvine, CA:
University of California, School of Information and Computer Science.
\url{https://archive.ics.uci.edu/}
}
\usage{
data(MSWeb)
}
\description{
Records the Vroots visited by users during a one-week period.
}
\details{
The data set was created by sampling and processing the www.microsoft.com logs.
It records site use by 38,000 anonymous, randomly selected users. For each user, it lists all areas of the web
site (Vroots) that the user visited during a one-week period in February 1998.

This data set contains 32,710 valid users and 285 Vroots.
}
\examples{
data(MSWeb)
MSWeb

nratings(MSWeb)

## look at first two users
as(MSWeb[1:2,], "list")

## items per user
hist(rowCounts(MSWeb), main="Distribution of Vroots visited per user")
}
\references{
J. Breese, D. Heckerman., C. Kadie (1998). Empirical Analysis of Predictive
Algorithms for Collaborative Filtering, Proceedings of the Fourteenth
Conference on Uncertainty in Artificial Intelligence, Madison, WI.
}
\seealso{
Other datasets:
\code{\link{Jester5k}},
\code{\link{MovieLense}}
}
\concept{datasets}
\keyword{datasets}
