% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\alias{plot,evaluationResults-method}
\alias{plot,evaluationResultList-method}
\title{Plot Evaluation Results}
\usage{
\S4method{plot}{evaluationResults}(x, y,
        avg = TRUE, add=FALSE, type= "b", annotate = FALSE, ...)
\S4method{plot}{evaluationResultList}(x, y,
        xlim=NULL, ylim=NULL, col = NULL, pch = NULL, lty = 1,
        avg = TRUE, type = "b", annotate= 0, legend="bottomright", ...)
}
\arguments{
\item{x}{the object to be plotted.}

\item{y}{a character string indicating the type of plot (e.g., "ROC" or "prec/rec").}

\item{avg}{plot average of runs?}

\item{add}{add to a plot?}

\item{type}{line type (see \code{plot}).}

\item{annotate}{annotate N (recommendation list size) to plot.}

\item{...}{further arguments passed on to \code{plot}.}

\item{xlim, ylim}{plot limits (see \code{plot}).}

\item{col}{colors (see \code{plot}).}

\item{pch}{point symbol to use (see \code{plot}).}

\item{lty}{line type (see \code{plot})}

\item{legend}{where to place legend (see \code{legend}).}
}
\description{
Creates precision-recall or ROC plots for recommender evaluation results.
}
\seealso{
\code{\linkS4class{evaluationResults}},
\code{\linkS4class{evaluationResultList}}.
See
\code{\link{evaluate}} for examples.

Other evaluation:
\code{\link[=calcPredictionAccuracy]{calcPredictionAccuracy()}},
\code{\link{error}},
\code{\link[=evaluate]{evaluate()}},
\code{\link{evaluationResultList-class}},
\code{\link{evaluationResults-class}},
\code{\link[=evaluationScheme]{evaluationScheme()}},
\code{\link{evaluationScheme-class}}
}
\concept{evaluation}
