% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{Recommender-class}
\alias{Recommender-class}
\alias{show,Recommender-method}
\title{Class "Recommender": A Recommender Model}
\description{
Represents a recommender model learned for a given data set
(a rating matrix).
}
\seealso{
See \code{\link{Recommender}} for the constructor function and
	a description of available methods.

Other recommender models:
\code{\link[=HybridRecommender]{HybridRecommender()}},
\code{\link[=Recommender]{Recommender()}}
}
\concept{recommender models}
\keyword{classes}
\section{Objects from the Class}{
Objects are created by
the creator function \code{Recommender(data, method, parameter = NULL)}
}
\section{Slots}{
	 \describe{
    \item{\code{method}:}{Object of class \code{"character"};
	used recommendation method.}
    \item{\code{dataType}:}{Object of class \code{"character"};
	concrete class of the input data.}
    \item{\code{ntrain}:}{Object of class \code{"integer"};
	size of training set. }
    \item{\code{model}:}{Object of class \code{"list"}; the model. }
    \item{\code{predict}:}{Object of class \code{"function"}; code to compute
	a recommendation using the model. }
  }
}
\section{Methods}{
  \describe{
    \item{getModel}{\code{signature(x = "Recommender")}: retrieve the model. }
    \item{predict}{\code{signature(object = "Recommender")}: create
	recommendations for new data (argument \code{newdata}). }
    \item{show}{\code{signature(object = "Recommender")} }
	 }
}
